[serve][llm] Address review: imports to top, return annotation, shared ingress utils
- pd_server: hoist function-local fastapi/starlette/ingress imports to module top.
- pd_server: annotate _pd_http_response return type as Response.
- Extract shared OpenAI ingress helpers into core/ingress/utils.py
  (NON_STREAMING_RESPONSE_TYPES, _sanitize_chat_completion_request,
  _peek_at_generator, _openai_json_wrapper, plus supporting
  _apply_openai_json_format and stream type aliases); ingress.py and
  pd_server now import them from utils.
- Revert the incidental host/port reorder in NixlConnectorBackend.setup.
